requirements.txt
pika
requests
fastapi
uvicorn[standard]
pytest
httpx
sqlalchemy
databases[asyncpg, aiosqlite]
python-dotenv
pydantic-settings
rich